requirements.txt
fastapi
uvicorn
supabase
python-dotenv
google-generativeai
httpx
pandas
pyarrow
scikit-learn
geopandas
shapely